[Inspector] Make Response tab faster (#180035)
## Summary

This PR improves the speed of "Response" tab rendering in the Inspector
flyout. The issue was that the heavy logic of building links ("Open in
Console", "Open in Search Profiler") was executed for Response tab too
although links can appear only for Request tab.

---------

(cherry picked from commit 7a00d2e24b38bbdb5b41c31fe0658d78067ee679)
