+++ src/plugins/inspector/public/views/requests/components/details/req_code_viewer.tsx
@@ -9,88 +9,26 @@
 // We want to allow both right-clicking to open in a new tab and clicking through
 // the "Open in Console" link. We could use `RedirectAppLinks` at the top level
 // but that inserts a div which messes up the layout of the inspector.
-/* eslint-disable @elastic/eui/href-or-on-click */
 
 import { EuiButtonEmpty, EuiCopy, EuiFlexGroup, EuiFlexItem, EuiSpacer } from '@elastic/eui';
-import type { ConnectionRequestParams } from '@elastic/transport';
 import { i18n } from '@kbn/i18n';
 import { XJsonLang } from '@kbn/monaco';
-import { compressToEncodedURIComponent } from 'lz-string';
-import React, { useCallback } from 'react';
-import { useKibana } from '@kbn/kibana-react-plugin/public';
+import React, { ReactNode } from 'react';
 import { CodeEditor } from '@kbn/code-editor';
-import { InspectorPluginStartDeps } from '../../../../plugin';
 
 interface RequestCodeViewerProps {
-  indexPattern?: string;
-  requestParams?: ConnectionRequestParams;
-  json: string;
+  value: string;
+  actions?: Array<{ name: string; action: ReactNode }>;
 }
 
 const copyToClipboardLabel = i18n.translate('inspector.requests.copyToClipboardLabel', {
   defaultMessage: 'Copy to clipboard',
 });
 
-const openInConsoleLabel = i18n.translate('inspector.requests.openInConsoleLabel', {
-  defaultMessage: 'Open in Console',
-});
-
-const openInSearchProfilerLabel = i18n.translate('inspector.requests.openInSearchProfilerLabel', {
-  defaultMessage: 'Open in Search Profiler',
-});
-
 /**
  * @internal
  */
-export const RequestCodeViewer = ({
-  indexPattern,
-  requestParams,
-  json,
-}: RequestCodeViewerProps) => {
-  const { services } = useKibana<InspectorPluginStartDeps>();
-
-  const navigateToUrl = services.application?.navigateToUrl;
-
-  function getValue() {
-    if (!requestParams) {
-      return json;
-    }
-
-    const fullPath = requestParams.querystring
-      ? `${requestParams.path}?${requestParams.querystring}`
-      : requestParams.path;
-
-    return `${requestParams.method} ${fullPath}\n${json}`;
-  }
-
-  const value = getValue();
-
-  const devToolsDataUri = compressToEncodedURIComponent(value);
-  const consoleHref = services.share.url.locators
-    .get('CONSOLE_APP_LOCATOR')
-    ?.useUrl({ loadFrom: `data:text/plain,${devToolsDataUri}` });
-  // Check if both the Dev Tools UI and the Console UI are enabled.
-  const canShowDevTools =
-    services.application?.capabilities?.dev_tools.show && consoleHref !== undefined;
-  const shouldShowDevToolsLink = !!(requestParams && canShowDevTools);
-  const handleDevToolsLinkClick = useCallback(
-    () => consoleHref && navigateToUrl && navigateToUrl(consoleHref),
-    [consoleHref, navigateToUrl]
-  );
-
-  const searchProfilerDataUri = compressToEncodedURIComponent(json);
-  const searchProfilerHref = services.share.url.locators
-    .get('SEARCH_PROFILER_LOCATOR')
-    ?.useUrl({ index: indexPattern, loadFrom: `data:text/plain,${searchProfilerDataUri}` });
-  // Check if both the Dev Tools UI and the SearchProfiler UI are enabled.
-  const canShowsearchProfiler =
-    services.application?.capabilities?.dev_tools.show && searchProfilerHref !== undefined;
-  const shouldShowsearchProfilerLink = !!(indexPattern && canShowsearchProfiler);
-  const handleSearchProfilerLinkClick = useCallback(
-    () => searchProfilerHref && navigateToUrl && navigateToUrl(searchProfilerHref),
-    [searchProfilerHref, navigateToUrl]
-  );
-
+export const RequestCodeViewer = ({ value, actions }: RequestCodeViewerProps) => {
   return (
     <EuiFlexGroup
       direction="column"
@@ -104,7 +42,7 @@ export const RequestCodeViewer = ({
         <EuiFlexGroup justifyContent="flexEnd" gutterSize="m" wrap>
           <EuiFlexItem grow={false}>
             <div>
-              <EuiCopy textToCopy={json}>
+              <EuiCopy textToCopy={value}>
                 {(copy) => (
                   <EuiButtonEmpty
                     size="xs"
@@ -119,38 +57,12 @@ export const RequestCodeViewer = ({
               </EuiCopy>
             </div>
           </EuiFlexItem>
-          {shouldShowDevToolsLink && (
-            <EuiFlexItem grow={false}>
-              <div>
-                <EuiButtonEmpty
-                  size="xs"
-                  flush="right"
-                  iconType="wrench"
-                  href={consoleHref}
-                  onClick={handleDevToolsLinkClick}
-                  data-test-subj="inspectorRequestOpenInConsoleButton"
-                >
-                  {openInConsoleLabel}
-                </EuiButtonEmpty>
-              </div>
-            </EuiFlexItem>
-          )}
-          {shouldShowsearchProfilerLink && (
-            <EuiFlexItem grow={false}>
-              <div>
-                <EuiButtonEmpty
-                  size="xs"
-                  flush="right"
-                  iconType="visBarHorizontal"
-                  href={searchProfilerHref}
-                  onClick={handleSearchProfilerLinkClick}
-                  data-test-subj="inspectorRequestOpenInSearchProfilerButton"
-                >
-                  {openInSearchProfilerLabel}
-                </EuiButtonEmpty>
-              </div>
-            </EuiFlexItem>
-          )}
+          {!!actions &&
+            actions.map((item) => (
+              <EuiFlexItem grow={false} key={item.name}>
+                <div>{item.action}</div>
+              </EuiFlexItem>
+            ))}
         </EuiFlexGroup>
       </EuiFlexItem>
       <EuiFlexItem grow={true} data-test-subj="inspectorRequestCodeViewerContainer">
