+++ src/plugins/inspector/public/views/requests/components/details/req_details_response.tsx
@@ -24,6 +24,6 @@ export class RequestDetailsResponse extends Component<DetailViewProps> {
       return null;
     }
 
-    return <RequestCodeViewer json={JSON.stringify(responseJSON, null, 2)} />;
+    return <RequestCodeViewer value={JSON.stringify(responseJSON, null, 2)} />;
   }
 }
