+++ src/plugins/inspector/public/views/requests/components/types.ts
@@ -6,8 +6,20 @@
  * Side Public License, v 1.
  */
 
+import type { SharePluginStart } from '@kbn/share-plugin/public';
+import type { CoreStart } from '@kbn/core-lifecycle-browser';
+
 export interface DetailViewData {
   name: string;
   label: string;
   component: any;
+}
+
+export interface InspectorKibanaServices {
+  share: SharePluginStart;
+  application: CoreStart['application'];
+  http: CoreStart['http'];
+  uiSettings: CoreStart['uiSettings'];
+  settings: CoreStart['settings'];
+  theme: CoreStart['theme'];
 }
